agents/personalization-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "personalization-agent"
version = "0.1.0"
description = "Personalization agent for multi-agent customer support system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "langgraph>=0.2.40",
    "langchain-aws>=0.2.2",
    "langchain>=0.3.7",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0"
]

[project.scripts]
personalization-agent = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]